+++ core/dataset.cpp
@@ -582,50 +582,6 @@ void dry_run_op(const DataProxy &a, const DataConstProxy &b, Op op) {
   dry_run::transform_in_place(a.data(), b.data(), op);
 }
 
-DataProxy DataProxy::operator+=(const DataConstProxy &other) const {
-  expect::coordsAndLabelsAreSuperset(*this, other);
-  data() += other.data();
-  return *this;
-}
-
-DataProxy DataProxy::operator-=(const DataConstProxy &other) const {
-  expect::coordsAndLabelsAreSuperset(*this, other);
-  data() -= other.data();
-  return *this;
-}
-
-DataProxy DataProxy::operator*=(const DataConstProxy &other) const {
-  expect::coordsAndLabelsAreSuperset(*this, other);
-  data() *= other.data();
-  return *this;
-}
-
-DataProxy DataProxy::operator/=(const DataConstProxy &other) const {
-  expect::coordsAndLabelsAreSuperset(*this, other);
-  data() /= other.data();
-  return *this;
-}
-
-DataProxy DataProxy::operator+=(const VariableConstProxy &other) const {
-  data() += other;
-  return *this;
-}
-
-DataProxy DataProxy::operator-=(const VariableConstProxy &other) const {
-  data() -= other;
-  return *this;
-}
-
-DataProxy DataProxy::operator*=(const VariableConstProxy &other) const {
-  data() *= other;
-  return *this;
-}
-
-DataProxy DataProxy::operator/=(const VariableConstProxy &other) const {
-  data() /= other;
-  return *this;
-}
-
 /// Return a const proxy to all coordinates of the dataset slice.
 ///
 /// This proxy includes only "dimension-coordinates". To access
@@ -794,185 +750,12 @@ std::unordered_map<Dim, scipp::index> DatasetConstProxy::dimensions() const {
   return base_dims;
 }
 
-constexpr static auto plus = [](const auto &a, const auto &b) { return a + b; };
-
-constexpr static auto minus = [](const auto &a, const auto &b) {
-  return a - b;
-};
-
-constexpr static auto times = [](const auto &a, const auto &b) {
-  return a * b;
-};
-
-constexpr static auto divide = [](const auto &a, const auto &b) {
-  return a / b;
-};
-
-template <class Op, class A, class B>
-auto &apply(const Op &op, A &a, const B &b) {
-  for (const auto & [ name, item ] : b)
-    dry_run_op(a[name], item, op);
-  for (const auto & [ name, item ] : b)
-    op(a[name], item);
-  return a;
-}
-
-template <class Op, class A, class B>
-decltype(auto) apply_with_delay(const Op &op, A &&a, const B &b) {
-  for (const auto &item : a)
-    dry_run_op(item.second, b, op);
-  // For `b` referencing data in `a` we delay operation. The alternative would
-  // be to make a deep copy of `other` before starting the iteration over items.
-  std::optional<DataProxy> delayed;
-  // Note the inefficiency here: We are comparing some or all of the coords and
-  // labels for each item. This could be improved by implementing the operations
-  // for detail::DatasetData instead of DataProxy.
-  for (const auto & [ name, item ] : a) {
-    static_cast<void>(name);
-    if (&item.underlying() == &b.underlying())
-      delayed = item;
-    else
-      op(item, b);
-  }
-  if (delayed)
-    op(*delayed, b);
-  return std::forward<A>(a);
-}
-
-template <class Op, class A, class B>
-auto apply_with_broadcast(const Op &op, const A &a, const B &b) {
-  Dataset res;
-  for (const auto & [ name, item ] : b)
-    if (const auto it = a.find(name); it != a.end())
-      res.setData(std::string(name), op(it->second, item));
-  return res;
-}
-
-template <class Op, class A>
-auto apply_with_broadcast(const Op &op, const A &a, const DataConstProxy &b) {
-  Dataset res;
-  for (const auto & [ name, item ] : a)
-    res.setData(std::string(name), op(item, b));
-  return res;
-}
-
-template <class Op, class B>
-auto apply_with_broadcast(const Op &op, const DataConstProxy &a, const B &b) {
-  Dataset res;
-  for (const auto & [ name, item ] : b)
-    res.setData(std::string(name), op(a, item));
-  return res;
-}
-
-[[nodiscard]] bool containsSparse(const DatasetConstProxy &ds) noexcept {
-  for (const auto &item : ds)
-    if (item.second.dims().sparse())
-      return true;
-  return false;
-}
-
-Dataset &Dataset::operator+=(const DataConstProxy &other) {
-  return apply_with_delay(operator_detail::plus_equals{}, *this, other);
-}
-
-Dataset &Dataset::operator-=(const DataConstProxy &other) {
-  return apply_with_delay(operator_detail::minus_equals{}, *this, other);
-}
-
-Dataset &Dataset::operator*=(const DataConstProxy &other) {
-  return apply_with_delay(operator_detail::times_equals{}, *this, other);
-}
-
-Dataset &Dataset::operator/=(const DataConstProxy &other) {
-  return apply_with_delay(operator_detail::divide_equals{}, *this, other);
-}
-
-Dataset &Dataset::operator+=(const DatasetConstProxy &other) {
-  return apply(operator_detail::plus_equals{}, *this, other);
-}
-
-Dataset &Dataset::operator-=(const DatasetConstProxy &other) {
-  return apply(operator_detail::minus_equals{}, *this, other);
-}
-
-Dataset &Dataset::operator*=(const DatasetConstProxy &other) {
-  return apply(operator_detail::times_equals{}, *this, other);
-}
-
-Dataset &Dataset::operator/=(const DatasetConstProxy &other) {
-  return apply(operator_detail::divide_equals{}, *this, other);
-}
-
-Dataset &Dataset::operator+=(const Dataset &other) {
-  return apply(operator_detail::plus_equals{}, *this, other);
-}
-
-Dataset &Dataset::operator-=(const Dataset &other) {
-  return apply(operator_detail::minus_equals{}, *this, other);
-}
-
-Dataset &Dataset::operator*=(const Dataset &other) {
-  return apply(operator_detail::times_equals{}, *this, other);
-}
-
-Dataset &Dataset::operator/=(const Dataset &other) {
-  return apply(operator_detail::divide_equals{}, *this, other);
-}
-
 std::unordered_map<Dim, scipp::index> Dataset::dimensions() const {
   std::unordered_map<Dim, scipp::index> all;
   for (const auto &dim : this->m_dims) {
     all[dim.first] = extents::decodeExtent(dim.second);
   }
   return all;
-}
-
-DatasetProxy DatasetProxy::operator+=(const DataConstProxy &other) const {
-  return apply_with_delay(operator_detail::plus_equals{}, *this, other);
-}
-
-DatasetProxy DatasetProxy::operator-=(const DataConstProxy &other) const {
-  return apply_with_delay(operator_detail::minus_equals{}, *this, other);
-}
-
-DatasetProxy DatasetProxy::operator*=(const DataConstProxy &other) const {
-  return apply_with_delay(operator_detail::times_equals{}, *this, other);
-}
-
-DatasetProxy DatasetProxy::operator/=(const DataConstProxy &other) const {
-  return apply_with_delay(operator_detail::divide_equals{}, *this, other);
-}
-
-DatasetProxy DatasetProxy::operator+=(const DatasetConstProxy &other) const {
-  return apply(operator_detail::plus_equals{}, *this, other);
-}
-
-DatasetProxy DatasetProxy::operator-=(const DatasetConstProxy &other) const {
-  return apply(operator_detail::minus_equals{}, *this, other);
-}
-
-DatasetProxy DatasetProxy::operator*=(const DatasetConstProxy &other) const {
-  return apply(operator_detail::times_equals{}, *this, other);
-}
-
-DatasetProxy DatasetProxy::operator/=(const DatasetConstProxy &other) const {
-  return apply(operator_detail::divide_equals{}, *this, other);
-}
-
-DatasetProxy DatasetProxy::operator+=(const Dataset &other) const {
-  return apply(operator_detail::plus_equals{}, *this, other);
-}
-
-DatasetProxy DatasetProxy::operator-=(const Dataset &other) const {
-  return apply(operator_detail::minus_equals{}, *this, other);
-}
-
-DatasetProxy DatasetProxy::operator*=(const Dataset &other) const {
-  return apply(operator_detail::times_equals{}, *this, other);
-}
-
-DatasetProxy DatasetProxy::operator/=(const Dataset &other) const {
-  return apply(operator_detail::divide_equals{}, *this, other);
 }
 
 std::ostream &operator<<(std::ostream &os, const DataConstProxy &data) {
@@ -1013,134 +796,6 @@ std::ostream &operator<<(std::ostream &os, const Variable &variable) {
 
 std::ostream &operator<<(std::ostream &os, const Dim dim) {
   return os << to_string(dim);
-}
-
-Dataset operator+(const Dataset &lhs, const Dataset &rhs) {
-  return apply_with_broadcast(plus, lhs, rhs);
-}
-
-Dataset operator+(const Dataset &lhs, const DatasetConstProxy &rhs) {
-  return apply_with_broadcast(plus, lhs, rhs);
-}
-
-Dataset operator+(const Dataset &lhs, const DataConstProxy &rhs) {
-  return apply_with_broadcast(plus, lhs, rhs);
-}
-
-Dataset operator+(const DatasetConstProxy &lhs, const Dataset &rhs) {
-  return apply_with_broadcast(plus, lhs, rhs);
-}
-
-Dataset operator+(const DatasetConstProxy &lhs, const DatasetConstProxy &rhs) {
-  return apply_with_broadcast(plus, lhs, rhs);
-}
-
-Dataset operator+(const DatasetConstProxy &lhs, const DataConstProxy &rhs) {
-  return apply_with_broadcast(plus, lhs, rhs);
-}
-
-Dataset operator+(const DataConstProxy &lhs, const Dataset &rhs) {
-  return apply_with_broadcast(plus, lhs, rhs);
-}
-
-Dataset operator+(const DataConstProxy &lhs, const DatasetConstProxy &rhs) {
-  return apply_with_broadcast(plus, lhs, rhs);
-}
-
-Dataset operator-(const Dataset &lhs, const Dataset &rhs) {
-  return apply_with_broadcast(minus, lhs, rhs);
-}
-
-Dataset operator-(const Dataset &lhs, const DatasetConstProxy &rhs) {
-  return apply_with_broadcast(minus, lhs, rhs);
-}
-
-Dataset operator-(const Dataset &lhs, const DataConstProxy &rhs) {
-  return apply_with_broadcast(minus, lhs, rhs);
-}
-
-Dataset operator-(const DatasetConstProxy &lhs, const Dataset &rhs) {
-  return apply_with_broadcast(minus, lhs, rhs);
-}
-
-Dataset operator-(const DatasetConstProxy &lhs, const DatasetConstProxy &rhs) {
-  return apply_with_broadcast(minus, lhs, rhs);
-}
-
-Dataset operator-(const DatasetConstProxy &lhs, const DataConstProxy &rhs) {
-  return apply_with_broadcast(minus, lhs, rhs);
-}
-
-Dataset operator-(const DataConstProxy &lhs, const Dataset &rhs) {
-  return apply_with_broadcast(minus, lhs, rhs);
-}
-
-Dataset operator-(const DataConstProxy &lhs, const DatasetConstProxy &rhs) {
-  return apply_with_broadcast(minus, lhs, rhs);
-}
-
-Dataset operator*(const Dataset &lhs, const Dataset &rhs) {
-  return apply_with_broadcast(times, lhs, rhs);
-}
-
-Dataset operator*(const Dataset &lhs, const DatasetConstProxy &rhs) {
-  return apply_with_broadcast(times, lhs, rhs);
-}
-
-Dataset operator*(const Dataset &lhs, const DataConstProxy &rhs) {
-  return apply_with_broadcast(times, lhs, rhs);
-}
-
-Dataset operator*(const DatasetConstProxy &lhs, const Dataset &rhs) {
-  return apply_with_broadcast(times, lhs, rhs);
-}
-
-Dataset operator*(const DatasetConstProxy &lhs, const DatasetConstProxy &rhs) {
-  return apply_with_broadcast(times, lhs, rhs);
-}
-
-Dataset operator*(const DatasetConstProxy &lhs, const DataConstProxy &rhs) {
-  return apply_with_broadcast(times, lhs, rhs);
-}
-
-Dataset operator*(const DataConstProxy &lhs, const Dataset &rhs) {
-  return apply_with_broadcast(times, lhs, rhs);
-}
-
-Dataset operator*(const DataConstProxy &lhs, const DatasetConstProxy &rhs) {
-  return apply_with_broadcast(times, lhs, rhs);
-}
-
-Dataset operator/(const Dataset &lhs, const Dataset &rhs) {
-  return apply_with_broadcast(divide, lhs, rhs);
-}
-
-Dataset operator/(const Dataset &lhs, const DatasetConstProxy &rhs) {
-  return apply_with_broadcast(divide, lhs, rhs);
-}
-
-Dataset operator/(const Dataset &lhs, const DataConstProxy &rhs) {
-  return apply_with_broadcast(divide, lhs, rhs);
-}
-
-Dataset operator/(const DatasetConstProxy &lhs, const Dataset &rhs) {
-  return apply_with_broadcast(divide, lhs, rhs);
-}
-
-Dataset operator/(const DatasetConstProxy &lhs, const DatasetConstProxy &rhs) {
-  return apply_with_broadcast(divide, lhs, rhs);
-}
-
-Dataset operator/(const DatasetConstProxy &lhs, const DataConstProxy &rhs) {
-  return apply_with_broadcast(divide, lhs, rhs);
-}
-
-Dataset operator/(const DataConstProxy &lhs, const Dataset &rhs) {
-  return apply_with_broadcast(divide, lhs, rhs);
-}
-
-Dataset operator/(const DataConstProxy &lhs, const DatasetConstProxy &rhs) {
-  return apply_with_broadcast(divide, lhs, rhs);
 }
 
 // For now this implementation is only for the simplest case of 2 dims (inner
